header att
- apagado a classe do header
- foi arrumado os botões do menu
